# back-end/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "API documentation for the application",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/Store/VnPay": {
            "post": {
                "tags": [
                    "VNPAY"
                ],
                "summary": "Thanh toán đơn hàng qua VNPAY",
                "description": "Tiến hành thanh toán đơn hàng qua cổng VNPAY, sau khi tạo đơn hàng thành công.",
                "operationId": "404c6ead59d5d32593917f7790711699",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "Mataikhoan",
                                    "PTTT",
                                    "chi_tiet"
                                ],
                                "properties": {
                                    "Mataikhoan": {
                                        "description": "Mã tài khoản của người đặt hàng",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "PTTT": {
                                        "description": "Phương thức thanh toán",
                                        "type": "string",
                                        "example": "Chuyển khoản"
                                    },
                                    "GhiChu": {
                                        "description": "Ghi chú cho đơn hàng",
                                        "type": "string",
                                        "example": "Ghi chú đơn hàng"
                                    },
                                    "Discount": {
                                        "description": "Giảm giá (nếu có)",
                                        "type": "number",
                                        "example": 10
                                    },
                                    "chi_tiet": {
                                        "description": "Chi tiết các sản phẩm trong đơn hàng",
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "MaSP": {
                                                    "description": "Mã sản phẩm",
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "SoLuong": {
                                                    "description": "Số lượng sản phẩm",
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Thanh toán thành công, chuyển hướng đến VNPAY",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thêm đơn hàng thành công"
                                        },
                                        "url": {
                                            "type": "string",
                                            "example": "https://sandbox.vnpayment.vn/paymentv2/vpcpay.html?..."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/OrderResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Lỗi xác thực hoặc dữ liệu không hợp lệ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thông báo lỗi chi tiết"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Lỗi xử lý trong quá trình thanh toán",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "fail"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Thanh toán thất bại"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "OrderResource": {
                "properties": {
                    "MaDonHang": {
                        "type": "integer",
                        "example": 1
                    },
                    "Mataikhoan": {
                        "type": "integer",
                        "example": 1
                    },
                    "TongTien": {
                        "type": "integer",
                        "example": 500000
                    },
                    "Discount": {
                        "type": "integer",
                        "example": 500000
                    },
                    "SoLuong": {
                        "type": "integer",
                        "example": 1
                    },
                    "Ten": {
                        "type": "string",
                        "example": "Nguyễn Văn A"
                    },
                    "SDT": {
                        "type": "string",
                        "example": "0123456789"
                    },
                    "DiaChi": {
                        "type": "string",
                        "example": "123 Đường ABC, Quận 1, TP.HCM"
                    },
                    "PTTT": {
                        "type": "string",
                        "example": "Chuyển khoản"
                    },
                    "GhiChu": {
                        "type": "string",
                        "example": "Ghi chú đơn hàng 1"
                    },
                    "Loai": {
                        "type": "string",
                        "example": "1"
                    },
                    "NgayDat": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-29T17:09:04Z"
                    },
                    "NgayGiao": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-11-01T17:09:04Z"
                    },
                    "TrangThai": {
                        "type": "string",
                        "example": "Đang xử lý"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-29T17:09:04Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-10-29T17:09:04Z"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}